fix remarks

+++ ErshovVladislav/Task4/Task4/src/main/java/ershov/ExamSystemController.java
@@ -32,7 +32,7 @@ public class ExamSystemController {
 	}
 
 	@RequestMapping("/count")
-	public int examSystemCount() {
+	public long examSystemCount() {
 		return examSystem.count();
 	}
