+++ ErshovVladislav/Task4/Task4/src/main/java/ershov/examsystem/IExamSystem.java
@@ -8,6 +8,6 @@ public interface IExamSystem {
 
     public boolean contains(long studentId, long courseId);
 
-    public int count();
+    public long count();
 
 }
